refactor(SafeValue): :recycle: move the InRangeTimeBased
